Fixed packaging error
