+++ spotbugsTestCases/src/java/multithreaded/sharedPrimitiveVariables/VolatileField.java
@@ -1,10 +1,9 @@
-package multithreaded.sharedPrimitiveVariables.sharedPrimitiveVariables;
+package multithreaded.sharedPrimitiveVariables;
 
-public class FieldWithBadVisibilityRunnable implements Runnable {
-    private boolean done = false;
+public class VolatileField {
+    private volatile boolean done = false;
 
-    @Override
-    public void run() {
+    public void doSomething() {
         while (!done) {
             try {
                 Thread.currentThread().sleep(1000);
